Move things around in crypto
